Rename CreatePost to PostManager and reduce the complexity
