+++ app/dashboard/create/page.tsx
@@ -1,12 +1,12 @@
 // local components
-import CreatePost from "@/app/components/createPost/CreatePost";
 import Dashboard from "@/app/components/dashboard/Dashboard";
+import PostManager from "@/app/components/postManager/PostManager";
 
 export default function Create() {
   return (
     <>
       <Dashboard>
-        <CreatePost />
+        <PostManager />
       </Dashboard>
     </>
   );
