+++ app/dashboard/post/edit/[id]/page.tsx
@@ -1,8 +1,8 @@
 "use client";
-import CreatePost from "@/app/components/createPost/CreatePost";
 import Dashboard from "@/app/components/dashboard/Dashboard";
 import useGetPosts from "@/app/hooks/useGetPosts";
 import { useParams } from "next/navigation";
+import PostManager from "@/app/components/postManager/PostManager";
 
 function PostEdit() {
   const { id } = useParams();
@@ -12,7 +12,7 @@ function PostEdit() {
   return (
     <>
       <Dashboard>
-        {post ? <CreatePost existingPost={post} /> : <p>Post not found</p>}
+        {post ? <PostManager existingPost={post} /> : <p>Post not found</p>}
       </Dashboard>
     </>
   );
